engine: finish extracting transpiled JS code into individual files
